feat(deident): 비식별 라이프사이클의 막힌 구간 셋을 열고 연동 주소를 설정으로 뺀다
R2 승인 영상 신고 차단 · R9 비식별 옵션 설정 · R3 신고 해소 시 산출물 선택 ·
R11 연동 서버 주소 설정 + 관리자 인증.

R2 — 검수 승인된 영상은 비식별 누락 신고를 접수하지 않는다(412, 역할 무관).
승인된 학습데이터는 되돌리지 않는다는 원칙이다. 이미 접수된 신고의 해소 경로는
그대로 열어 둔다 — 막으면 그 영상이 작업락 + DE_IDENT_YN='F' 로 영구 고착된다.
⚠ 인지·수용한 잔여 위험: 승인 후 잔존 개인정보를 발견해도 신고·재비식별 어느
쪽도 열리지 않는다(재비식별은 별도 사유로 이미 409). 발견은 승인 전이 전제다.

R9 — 마스킹 방식·범위·프레임 저장 3종을 설정 화면에서 조정한다.
masking_range 가 int 라 0.5 가 0 으로 잘려 애초에 전송이 불가능하던 선행 결함을
실수로 교정했다. masking_type 은 벤더가 1을 할당하지 않아 범위가 아니라
허용값 집합{0,2,3}으로 검증한다. exp_quality·exp_format 은 벤더 미지원이라
화면에 노출하지 않되 규격상 필수라 요청에는 계속 싣는다.

R3 — 신고 해소 시 재비식별 산출물을 목록에서 고른다. 외부 솔루션은 같은 이름으로
덮어쓰지 않고 다른 이름으로 산출하므로(KPST 는 {stem}-mask{ext}), 원장 경로 하나의
mtime 만 보던 기존 판정으로는 그런 신고가 영영 해소되지 않았다. 제출값은 서버가
만든 목록과의 대조로만 수락해 경로를 조립하지 않는다. 선택 결과를 원장에 새 행으로
적재해야 프레임 재추출·스트리밍이 새 파일을 쓴다 — 이게 없으면 목록 선택이 반쪽이다.

R11 — 연동 4종 주소를 설정으로 빼고 호출 시점에 다시 읽는다(재기동 불필요).
비식별 키는 kpst.deid.base-url 이다 — 구 후보가 구동하는 빈은 주입처가 0건이라
그 키로는 아무것도 바뀌지 않았다. 주소 대역 차단은 하지 않는다(온프렘 내부망 배포 ·
대상이 내부 GPU 서버 · 아웃바운드는 인프라 계층이 통제). 관리자 인증은 기존 공유
패스워드를 재사용한 무상태 서명 토큰이며 기본 10분 · 상한 30분이고 역할을
승격시키지 않는다. 감사는 새 테이블 없이 기존 컬럼 + 로그로 남긴다.

보안 심층 리뷰의 HIGH 2건을 함께 닫았다 — 목/시뮬레이터 비식별 서버를 prd 에서
막던 가드가 런타임 override 로 조용히 우회되던 것(축이 IP 대역이 아니라 호스트명이라
대역 차단 폐기와 별개다), 주소를 바꾸면 정적 토큰이 새 호스트로 그대로 따라가던 것.

검증: FULL 회귀 PASS — backend 6,171 / frontend 2,880 / 실패 0 / contrastGuard 94.

+++ backend/src/main/java/kr/co/cudo/authoring/common/client/dto/KpstProjectRequest.java
@@ -22,14 +22,18 @@ public record KpstProjectRequest(
         List<String> files,
         @JsonProperty("masking_type") int maskingType,
         @JsonProperty("db_save") int dbSave,
-        @JsonProperty("masking_range") int maskingRange,
+        @JsonProperty("masking_range") double maskingRange,
         @JsonProperty("exp_quality") int expQuality,
         @JsonProperty("exp_format") int expFormat
 ) {
     /** 규격 기본값 (§22.4 부록 A). 매직 넘버 금지 — 호출자 미지정 시 사용. */
     public static final int DEFAULT_MASKING_TYPE = 0;
     public static final int DEFAULT_DB_SAVE = 0;
-    public static final int DEFAULT_MASKING_RANGE = 1;
+    /**
+     * 마스킹 영역 배율 기본값. <b>실수</b>(0.5~2.0)다 — 벤더 확인 결과 이 필드는 코드값이 아니라
+     * 배율이며, {@code int} 로 두면 0.5 가 0 으로 잘려 전송 자체가 불가능하다(선행 결함 교정).
+     */
+    public static final double DEFAULT_MASKING_RANGE = 1.0;
     public static final int DEFAULT_EXP_QUALITY = 0;
     public static final int DEFAULT_EXP_FORMAT = 1;
